cli arguments with animation input
