+++ Cargo.toml
@@ -18,6 +18,7 @@ image = "0.24"
 tempfile = "3"
 tobj = "4"
 regex = "1"
+async-trait = "0.1.88"
 
 [dependencies.gltf]
 version = "1.4"
