--- src/scenes/mod.rs
@@ -1,3 +1,0 @@
-pub mod obj;
-pub mod custom_scene;
-pub mod required_scene;
